chore: rename trpc to backend and remove mentions of trpc
