+++ apps/web/src/components/forms/recipe/RecipeForm.tsx
@@ -23,7 +23,7 @@ import {
 } from '@/components/ui/sortable';
 import { Textarea } from '@/components/ui/textarea';
 import { catchError } from '@/lib/utils';
-import { createRecipeInput } from '@/trpc/recipe/test';
+import { createRecipeInput } from '@/backend/recipe/test';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { DragHandleDots2Icon } from '@radix-ui/react-icons';
 import * as React from 'react';
