+++ apps/web/src/components/forms/recipe/UpdateRecipeForm.tsx
@@ -2,7 +2,7 @@
 
 import { updateRecipeRevalidate } from '@/app/actions';
 import { catchError } from '@/lib/utils';
-import { updateRecipe } from '@/trpc/recipe/recipeRouter';
+import { updateRecipe } from '@/backend/recipe/recipeRouter';
 import { useRouter } from 'next/navigation';
 import { toast } from 'sonner';
 import { RecipeForm, type RecipeFormInput } from './RecipeForm';
