+++ apps/web/src/components/recipeView.tsx
@@ -6,8 +6,8 @@ import { Shell } from '@/components/shells/shell';
 import { Checkbox } from '@/components/ui/checkbox';
 import { Separator } from '@/components/ui/separator';
 import { cn } from '@/lib/utils';
-import type { getPublicRecipe } from '@/trpc/recipe/recipeRouter';
-import { deleteRecipe } from '@/trpc/recipe/recipeRouter';
+import type { getPublicRecipe } from '@/backend/recipe/recipeRouter';
+import { deleteRecipe } from '@/backend/recipe/recipeRouter';
 import { type RecipeRating } from '@prisma/client';
 import { Tooltip } from '@radix-ui/react-tooltip';
 import { useAction } from 'next-safe-action/hooks';
