+++ apps/web/src/content/pages/about.mdx
@@ -12,7 +12,6 @@ The project is still in development. You can follow the progress on [GitiHub](ht
 - [Tailwind CSS](https://tailwindcss.com)
 - [Clerk Auth](https://clerk.dev)
 - [Prisma ORM](https://www.prisma.io/)
-- [TRPC](https://trpc.io)
 - [Contentlayer](https://www.contentlayer.dev)
 - [uploadthing](https://uploadthing.com)
 - [Stripe](https://stripe.com)
@@ -20,7 +19,6 @@ The project is still in development. You can follow the progress on [GitiHub](ht
 ## Credits
 
 - [SkateShop](https://skateshop.sadmn.com/) - For E-commerce, and stripe account connection for each store
-- [Quill](https://github.com/joschan21/quill) - A Modern SaaS-Platform Built With Next.js 13 using TRPC and Prisma
 - [Acme Corp](https://acme-corp.jumr.dev) - For the OAuth implementation with Clerk, and AuthLayout
 - [craft.mxkaske.dev](https://craft.mxkaske.dev) - For the awesome fancy components
 - [Taxonomy](https://tx.shadcn.com) - For contentlayer, and stripe subscription implementation
